requirements.txt
torch>=1.9.0
torchvision
timm
Pillow
psutil
GPUtil
ultralytics